+++ src/test/java/com/braintreegateway/integrationtest/BraintreeGatewayIT.java
@@ -4,16 +4,16 @@ import com.braintreegateway.BraintreeGateway;
 import com.braintreegateway.Environment;
 import com.braintreegateway.Configuration;
 import com.braintreegateway.testhelpers.TestHelper;
-import org.junit.Test;
 
+import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 public class BraintreeGatewayIT {
+
     @Test
     public void developmentBaseMerchantUrl() {
-        BraintreeGateway gateway = new BraintreeGateway(Environment.DEVELOPMENT, "integration_merchant_id", "publicKey",
-                "privateKey");
+        BraintreeGateway gateway = new BraintreeGateway(Environment.DEVELOPMENT, "integration_merchant_id", "publicKey", "privateKey");
         String port = System.getenv().get("GATEWAY_PORT") == null ? "3000" : System.getenv().get("GATEWAY_PORT");
         Configuration configuration = gateway.getConfiguration();
         assertEquals("http://localhost:" + port + "/merchants/integration_merchant_id", configuration.getBaseURL() + configuration.getMerchantPath());
@@ -39,12 +39,4 @@ public class BraintreeGatewayIT {
         Configuration configuration = gateway.getConfiguration();
         assertEquals("https://api.braintreegateway.com:443/merchants/production_merchant_id", configuration.getBaseURL() + configuration.getMerchantPath());
     }
-
-    @Test
-    public void partnerIdAlias() {
-        BraintreeGateway gateway = BraintreeGateway.forPartner(Environment.DEVELOPMENT, "partner_id", "publicKey", "privateKey");
-        assertEquals(gateway.getConfiguration().getPublicKey(), "publicKey");
-        assertEquals(gateway.getConfiguration().getPrivateKey(), "privateKey");
-    }
-
 }