+++ src/test/java/com/braintreegateway/integrationtest/CustomerIT.java
@@ -159,6 +159,22 @@ public class CustomerIT extends IntegrationTest {
     }
 
     @Test
+    public void createWithRiskDataParams() {
+        CustomerRequest request = new CustomerRequest().
+            creditCard().
+                number("4111111111111111").
+                expirationDate("05/12").
+                done().
+            riskData().
+                customerBrowser("IE6").
+                customerIP("192.168.0.1").
+                done();
+        Result<Customer> result = gateway.customer().create(request);
+
+        assertTrue(result.isSuccess());
+    }
+
+    @Test
     public void createWithCustomFields() {
         CustomerRequest request = new CustomerRequest().
             customField("store_me", "custom value").
@@ -795,6 +811,87 @@ public class CustomerIT extends IntegrationTest {
         assertEquals("555-555-5555", updatedCustomer.getFax());
         assertEquals("555-555-5554", updatedCustomer.getPhone());
         assertEquals("http://getbraintree.com", updatedCustomer.getWebsite());
+    }
+
+    @Test
+    public void updateDefaultPaymentMethodInOptions() {
+        CustomerRequest request = new CustomerRequest().
+            firstName("Mark").
+            lastName("Jones");
+
+        Customer customer = gateway.customer().create(request).getTarget();
+
+        String token1 = "TOKEN-" + new Random().nextInt();
+
+        PaymentMethodRequest request1 = new PaymentMethodRequest().
+            paymentMethodNonce(Nonce.TransactableVisa).
+            token(token1).
+            customerId(customer.getId());
+
+        Result<? extends PaymentMethod> result1 = gateway.paymentMethod().create(request1);
+        assertTrue(result1.getTarget().isDefault());
+
+        String token2 = "TOKEN-" + new Random().nextInt();
+
+        PaymentMethodRequest request2 = new PaymentMethodRequest().
+            paymentMethodNonce(Nonce.TransactableMasterCard).
+            token(token2).
+            customerId(customer.getId());
+
+        gateway.paymentMethod().create(request2);
+        PaymentMethod newPaymentMethod = gateway.paymentMethod().find(token2);
+        assertFalse(newPaymentMethod.isDefault());
+
+        CustomerRequest updateRequest = new CustomerRequest().
+            creditCard().
+                options().
+                    updateExistingToken(token2).
+                    makeDefault(true).
+                    done().
+                done();
+
+        gateway.customer().update(customer.getId(), updateRequest).getTarget();
+
+        newPaymentMethod = gateway.paymentMethod().find(token2);
+        assertTrue(newPaymentMethod.isDefault());
+    }
+
+    @Test
+    public void updateDefaultPaymentMethod() {
+        CustomerRequest request = new CustomerRequest().
+            firstName("Mark").
+            lastName("Jones");
+
+        Customer customer = gateway.customer().create(request).getTarget();
+
+        String token1 = "TOKEN-" + new Random().nextInt();
+
+        PaymentMethodRequest request1 = new PaymentMethodRequest().
+            paymentMethodNonce(Nonce.TransactableVisa).
+            token(token1).
+            customerId(customer.getId());
+
+        Result<? extends PaymentMethod> result1 = gateway.paymentMethod().create(request1);
+        assertTrue(result1.getTarget().isDefault());
+
+        String token2 = "TOKEN-" + new Random().nextInt();
+
+        PaymentMethodRequest request2 = new PaymentMethodRequest().
+            paymentMethodNonce(Nonce.TransactableMasterCard).
+            token(token2).
+            customerId(customer.getId());
+
+        gateway.paymentMethod().create(request2);
+        PaymentMethod newPaymentMethod = gateway.paymentMethod().find(token2);
+        assertFalse(newPaymentMethod.isDefault());
+
+        CustomerRequest updateRequest = new CustomerRequest().
+            defaultPaymentMethodToken(token2);
+
+        gateway.customer().update(customer.getId(), updateRequest).getTarget();
+
+        newPaymentMethod = gateway.paymentMethod().find(token2);
+        assertTrue(newPaymentMethod.isDefault());
     }
 
     @Test